cmd/thor/pruner: rename pkg optimizer to pruner
