+++ cmd/thor/pruner/optimizer_test.go
